requirements.txt
streamlit
torch
torchvision
numpy
Pillow
opencv-python
scikit-image